Simplify default usage of clang-tidy-cache (#3)
* Simplify default usage of clang-tidy-cache

Currently clang-tidy-cache requires some non-default arguments to function.
This makes it a bit harder to start using, as for example without
-export-fixes it simply exited with exit code 1 and no error message.

To fix that, we make -export-fixes optional and add support for alternative
command line specification of the export fixes file. We also make the build
root an optional argument, deriving if from the linted filename the same
way clang-tidy does. Note that this only works correctly if just one file
is used. We also print error messages to assist the user figuring out the
cause of non-zero exit.

We also add support for both -p <path> and -p=<path> syntax for the build
root, as clang-tidy supports both.

We also refactor the cache storing, since export fixes is now optional. We
explicitly pass the contents to store to the cache functions, and let the
caller decide what to do with the contents (currently it's storing the
exports file).

* Apply go fmt

* Add go.mod for easier usage.

* Add support for overriding cache directory

Add support for environment variable CLANG_TIDY_CACHE_DIR (inspired by CCACHE_DIR), which allows overriding the directory where the cache is stored. This makes it easier to control the cache from CI environments.

* Improve support for compile_commands.json search

Clang-tidy searches compile_commands.json in the expected location or in any of the parent directories. We add the same logic to clang-tidy-cache, so that it can be easier used with for example CMake builds with nested directories but just a single top-level compile_commands.json file.