Provide a default clang-tidy path and env variable override
It's simpler for users to get started if they don't need to create
a config file at all, so this commit provides a default: simply use
the `clang-tidy` that's on the path.

It's also often useful to be able to override the config via env
variables without touching the config file: `CLANG_TIDY_CACHE_BINARY`
makes that possible.