Add basedir config to reuse cache from different directories
This feature is equivalent to `CCACHE_BASEDIR` and enables cache reuse
across directories. This is needed because some absolute paths can end
up in the preprocessed output via `__FILE__` macros or other debugging
mechanisms. This causes cache misses even though the code is otherwise
identical.

When `CLANG_TIDY_CACHE_BASEDIR`/`base_dir` is set, we replace absolute
paths that match that base prefix with relative paths in the preprocessed
output.

Additionally, we need to pass the `-P` flag to the preprocessor to avoid
linemarker comments that indicate which bit of the output came from which
header path.